Add ability to restart after partial regridding failure
This commit resolves #6 for addressing a partial regridding of the source
dataset.  As a default, running regrid.py will check for variables in
a pre-existing zarr file.  If those variables exist and the chunks are
fully initialized, then the variable is skipped.  If the variable is not
fully initialized or doesn't exist the full variable will be regridded.
Also adds a flag (-f or --force_regrid) for regrid.py in case you want to
force the entire dataset to be regridded