Add new 15-minute 1-deg data to intake and updated dataset loading
Intake now has the information for the 1-degree regridded 15-minute
SHiELD data.  The open_dataset scheme was set up in such a way that
it required code changes for new tags to access the specified datasets.
Now we can specify shortened dataset names in `short_datatag_defs.yml`
and connect those to dataset keys described in `catalog.yml`.
`src.data.open_dataset()` checks for a short_tag and uses the catalog
key specified or, if no short tag is present, will use the specified
key directly in the catalog, raising a KeyError if no entry is found.

This commit resolves #5.