Specify data transforms in the catalog file
For datasets we may need to apply certain transformations to prepare
them for use such as variable renaming or coordinate adjustments.
This commit enables a new tag that can be specified in the metadata
section of the the intake catalog for a datasource, "data_tranforms".
The list of transform function names (assumed to take one argument, an
xarray dataset, and returns the same) are applied to the dataset
after it is opened from the catalog.  If the function requested does not
exist in src.data, a KeyError is raised.