Run snakemake within docker (#19)
* Add dockerfile

* Use requirements.txt for pip

* install gcloud sdk in the docker image

* Use compiled version of crcmod for gsutil

without this the gsutil commands in the snakemake workflow will
fail.

this required a particularly arcane combination of pip and
apt-get commands. In particular, python-dev with pip must be installed.

* Change run rule to work within docker image

This detects if the fv3.exe file is present at /FV3/fv3.exe, and runs it
using subprocess directly if so.

* Fix errors with GCS files

Sometimes snakemake would mysteriously fail when downloading the
orographic data. Changing the keep_local flag seemed to fix this.

* Fix ability to run workkflow outside of docker

A previous commit broke this because it saved the submit_job.sh
to the wrong location.

* upload completed restart directory to GCS

* Change name

* Add scripts to provision k8s cluster

* Package workflow as dockerfile

To be orchestrated by kubernetes this workflow should contain
all of its component pieces

* Add instructions for using argo to the readme

* Update README

* Add argo workflow looping over inputs

* Provision nodes with ssd, might be faster

* run argo workflow for several time steps

* Add argo rule to run just the Fv3 step

* Add memory limit to argo run

* Document machine-type requirements for run

* Move argo scripts to argo folder for cleanliness

Also use a new image for the analysis

* Restore location of install_gcloud script

* Build image in circleci

* Fix docker image building

* Add sample workflow directory

* Move installation step to beginning of Dockerfile

Previously, it had to recompile the image every time

* Move gcloud application default credentials into docker

* Fix run make command

* Setup remote docker

* Disable docker build since this requires authentication