Automate submit sfccoarsen job (#45)
This commit contains some packaging improvements and adds some rules to Makefiles, which simplifies running this job and setting up the software environment required to do so. @frodre I think it might be nice to add a similar Makefile to the extraction workflow directory.

I also fixed a bug where the surface coarse-graining is run with multiple dask threads, which fails dramatically when scaled in Dataflow. I fixed this in previous commits, but somehow it got reintroduced.

Changes:
* add makefile for submitting the coarening dataflow job
* fix typo
* Add logging statement to url checking routine
* Fix package of dataflow
* Must use single-threaded scheduler for dataflow