Coarsen surface diagnostics from C3072 SHIELD run (#64)
This PR adds a dataflow pipeline and corresponding tests for coarse-graining the C3072 SHIELD data. It also includes some common apache beam functionality (in fv3net.common) that could be shared across different pipelines.

The old draft PR (#60) was difficult to merge with the great-refactor, so I just moved the relevant files over to this new branch. I will close the other ones.

Todo

 add documentation to the common.py module.
 close #60
 delete the various old branches.
 manually checked that the dataflow job still works.
 tests pass

Changes include:

* Add surface coarening pipeline

This is the only code contributed by the surface coarsening branch

* Add test

* Fix import errors

* Run automatic formatter and add readme

* Update readme and add local script

* Document the routines in common.py

* Add nose to requirements

This is needed by the apache_beam TestPipeline object.

* Finally get a working combination of sdist and setup.py

* Simplify and comment the code

Removes the netCDF sink class, which was confusing.

* Add type hint

* fix broken link in docstring

* Update workflows/coarsen_3072_surface_diagnostics/submit_job.sh

Co-Authored-By: Oliver Watt-Meyer <oliwm@vulcan.com>

* remove stupid line