Add function to interpolate unstructured data (#96)
This commit add a function to interpolate xarray data on an unstructured grid using nearest neighbors. This makes up for a shortcomming in xarray's interp, which won't work if the coordinates are multidimensional. I have been using to interpolate onto cross sections (e.g. lat=0, lon=1-360), and the code is very fast.