Workflow to convert diagnostic output to zarr store (#123)
This workflow takes a path/url to a run directory and converts all files of the form {prefix}.tile*.nc in the root of the run directory (that is, all the diagnostic outputs) to zarr stores. One zarr store is created for each category of diagnostic output. By default the zarr stores are created in the parent of the run directory. Optionally, the specific category or categories to convert and the location of the output zarr store can be specified.

* Add pipeline for netCDF to zarr diagnostics conversion

* Add dataflow submission script

* Uncomment some diagnostic categories

* Determine diagnostic categories if none provided

* Use fsspec instead of gsutil to copy zarr

* Add some logging statements

* Add README and workflow-specific setup.py

* Pass rundir as argument to submit_job.sh

* Update README.md

* Add uuid to dataflow job name

* Increase CPU count for dataflow job

* Change chunk size and copy zarr with gsutil

* Use bigger worker, pass vcm extra package

* Update README

* Add consolidate_metadata step