Feature/prognostic run (#113)
This PR introduces an example of how to run the model with an scikit learn model trained by @AnnaKwa. It includes

a runfile for fv3gfs-python which loads a scikit-learn model and uses it to make a prediction every timestep.
A dockerfile to recreate the environment
Add a module state_io.py for performing parallel I/O to zarr and pickle files from the python time-loop. This is used to save a file test.zarr of the python-level diagnostics (e.g. net_precipitation, heating, PW)
A script plot_run.py for making movies of the fields
A jupyter notebook for plotting some global averages