Add workflow for single FV3GFS runs, free or nudged (#83)
This PR adds a workflow that submits one off FV3GFS runs. It can handle both free-running and nudged simulations. The job submission script is re-usable, and specific runs (both in terms of model configuration, and in terms of CPU/memory required) are specified using yaml files. The location of output is specified as a command-line argument.