Merge with one-step refactoring
The original branch did not include the latest one-step updates from the
master branch.  Merged in but it conflicted with some of the previous
refactorings of submit_jobs into submit_utils.  Re-did the utils
functions with the latest updates adhering to no low-level helper
function directory manipulations.  This makes it easier for
orchestration to provide input/output information without assumption of
what's happening in the workflow step.