Slightly optimize open_restarts (#162)
This PR introduces a slight optimization that is useful for open_restarts.

Previously combine_array_sequence made many calls to xarray concat when combining each dimension. Now, it collects all the data for a given dimension before concatenating. This is faster because there is only one concat call per dimension to combine.

The tests for combine_array_sequence are still passing.