Make regrid_vertical dask-compatible (#202)
Modify regrid_vertical to make it dask-compatible. For now I've kept the old version of regrid_vertical and named it regrid_vertical_legacy just to illustrate that the new version produces identical outputs. To test that the dask version works, I've tested that when passed dask arrays, results remain lazy, and when computed they match the results when passed NumPy arrays.

I've taken care to add code to adjust the chunks of all the dask inputs to make sure that they are compatible with the way the dask version of the function works (applied to sets of columns at a time through dask.array.map_blocks). This way the user doesn't necessarily need to worry about that ahead of time.

With these modifications, I'm able to coarse-grain the 161 GB C3072 fv_tracer and 108 GB C3072 fv_core restart files on pressure in an out-of-core fashion on my VM, which has ~15 GB of RAM.