Remove run_tests in image
Running the tests in the image is no longer necessary and was testing
external/fv3config for some reason, even though that folder is excluded
in pytest.ini. I don't think this second round of tests is needed, since
we have run pytest before, and we have also run the integration tests.