Resolve issue with NaNs in the training pipeline
Because there are no diagnostics saved for the initial condition
step = 'begin' and forecast_time=0 will have no data for the surface
quanties. This was somewhat by design

This PR instead grabs the one-step data from after the physics step,
where there is data for all forecast times. This is more consistent with
where we compute the forcings anyway.