Fix integration tests again (#248)
The tests are broken because the wrong yamls were used by the create_training_data step. This PR updates those configurations and uses configmaps to deploy them into the end-to-end pipeline, rather than using the configurations backed into the docker image. This allows us to more easily retest the same image with different configurations.

Other changes to the ML training and training data pipelines are
- grab data from step="after_dynamics" to avoid NaNs
- remove many try-catch blocks. We should only be catching I/O errors, not code/input data that we are responsible for.


This is a quick fix until we merge #250.

Resolves to  #247