One-step diags report workflow (#206)
This PR introduces a micropackage workflow step ("one_step_diags") that processes output from the one-step runs into a set of figures in a report. The goal is to understand the behavior of the one-step runs relative to the hi-res coarsened data across their forecast evolution. 

The workflow step is divided up into a preprocessing step (via dataflow), which dumps a netcdf of processed data need for the plots to gcs; that netcdf is then used by a plotting routine to make the report.