Synthetic data regression tests (#298)
Introduces a regression testing framework for testing workflow steps with synthetic data. This will reduce the burden on the end-to-end pipeline and allow us to easily mock the inputs for individual workflow steps for regression tests. Changes include:

* A new package `external/synth` for creating "schema" from existing zarr datasets, saving schema to disk, and generating random xarray data from the "schema"
* Add documentation in `docs/regression-testing.rst`
* Use synthetic data to test the create-training data pipeline. These tests run in 30 seconds with synthetic data.
* Hook the regression tests into Circle CI

Overall, this is a good pattern for future regression tests using xarray-like data. To use this, either some minor refactoring needs to be done, so that the tested components can take in "xarray" data. Or the synthetic data can be saved to disk/cloud storage and used with the existing CLI.

Remaining Todos: 

- [x] Hook into CircleCI
- [x] Test create training data pipeline at scale to ensure it still works.

Resolves #267