Fix pooch parallel deployment bug (#326)
A package (pooch) used by some of our dependencies (metpy and scikit-image) was causing crashes on dataflow and k8s because it wasn't properly supporting multiprocess/parallel deployments with shared filesystems (#323 ). A recent new release from pooch appears to solve that problem based on discussions there (scikit-image/scikit-image#4660), and on some brief testing by @brianhenn and @mcgibbon on our jobs. This one was hard to reproduce because it is dependent on random concurrency but it appears to have been solved by using pooch release 1.1.1.

So this PR pins our fv3net conda env and the fv3net and vcm setup required packages to pooch==1.1.1.