Add script to check for large files (#338)
We have accidentally checked in two 7 MB notebook files with embedded pngs to the fv3net repository. This commit adds a CI hook that will prevent merging commits with large files.

Adds a script that will prevent large files in the commit. This script has a  threshold (currently set at 250 Kb) and a whitelist that can be easily added to. It will be triggered when running `make lint`. 

Here is an example of what the CircleCI output:

<img width="748" alt="Screen Shot 2020-05-16 at 3 29 21 PM" src="https://user-images.githubusercontent.com/1386642/82131456-39400780-978a-11ea-8dcc-2887019c7710.png">


Resolves #314