Refactor to functional API for loading batches (#307)
This refactor PR includes a change to use selectable functions to load batches.

We want to add the option to support multiple data sources for the same training step, and be able to modify only configuration to get a different data source to be used. This PR adds an interface for specifying a function to load batches. The function can be custom-written for specific data sources. In addition, I've added the option to rename variables on-disk to more standard names used for model inputs and outputs. This can be changed within the configuration file, without changing functions.

In this PR, a renaming functionality is added to map data storage names to standard names, but using it currently breaks later pipeline steps (the diagnostic evaluation). Fixing the later steps will be done in a later PR. For now the integration tests do not use this feature.

Major changes:
- BatchGenerator is now removed in favor of functions which return a Sequence object of batches. In this PR, a "Batch" is still an xarray dataset. This will be changed in a future PR.
- train data location is no longer present in the ModelTrainingConfig, since it is not present in the yaml configuration file. Instead it is passed as its own argument to the batch generation routine.
- ModelTrainingConfig is no longer responsible for manipulating configuration, only for containing it.
- Options used for batch generation are moved to a batch_kwargs subtree in the yaml configuration and in ModelTrainingConfig, so that different batch generation functions can have different call signatures
- The model now can use long names for e.g. total sky downward shortwave flux at top of atmosphere, with logic for renaming the variable on disk present in the configuration yaml
- "none", "None", and None are no longer valid surface types for mask_to_surface_type. The "global" surface type is still available for this option.
- the sklearn regression workflow now only has one configuration file in the repo, in the end-to-end workflow (it must be there due to kustomization requirements)
- Batches will now each train using a different (deterministic) random seed.

Minor changes:
- some imports in fv3net changed from `from submodule import names` to `import submodule`
- In the regression workflow, it is now possible to set n_jobs manually as a hyperparameter on the random forest regressor
- argument parsing for regression workflow is contained in its own function in main