Version synth schema and refactor "range" concept (#329)
Generating schema was needlessly complicated by the need to define a "sampling" function. This difficulty because "Range" was an attribute of the `VariableSchema`, a poor separation of concerns.

Refactor the range and generation functions to not be owned by the `Schema` objects. If they choose too, the user can specify the ranges they care about, otherwise a default is used.

Other changes include:
- Version the python dictionary representation of the "DatasetSchema"  to ensure backwards compatibility when reading older test
- Refactor `generate` functionality out of the schema classes. This functionality should not be tightly bound to these data objects since it can change independent reasons.