Compute Q1/Q2 from C3072 budget (single-threaded) (#313)
The one-step method for computing Q1 and Q2 is prone to initialization shocks over many time-scales. An alternative method is to compute this apparent sources by a budget analysis of the fine-res model. This commit introduces an apache beam pipeline for doing this.

Changes include:
- Generalize the implementation of `vcm.round_time`
- Refactor the well-tested `dump_nc` tool from the prognostic run diagnostics to `vcm`
- Import some previously private functions in `vcm.cubedsphere`
- Add a module `workflows/fine_res_budget` containing this pipeline
   - includes tests (100% coverage of new feature :=) )
   - usage documented with brief README

Note that this pipeline has only been tested locally, since remote execution with Google Cloud 
 Dataflow is not currently working (see #321 ). A follow-up commit will add that feature.