Refactor diurnal sample report to use time as any other variable (#342)
In reviewing #307, @frodre pointed out it is awkward to return a tuple with this special time_list object, especially when depending on how the data is stored, this might not be available, and depending on how the data is used, it might not be required. 

This PR changes that code to use the coordinate variable stored in the batch datasets. It also standardizes the name of the time dimension to a module constant.

Major changes:
- FunctionOutputSequence now takes in a sequence of arguments, instead of a sequence of sequences of arguments. This simplifies the code that uses it quite a bit, especially for some upcoming plans to transform sequences.
- time dimension in batches now uses name fv3util.regression.TIME_NAME