Enhance reliability of restart coarsening (#369)
This pipeline previously struggled to complete without failures. Because it wasn't very modular, these errors are not easy to guard with retry logic. A good deal of this complexity involved with expliclty downloading data to the local file system and associated path-naming logic.

# Changes to coaren_restarts pipeline

This PR is basically a re-write of this pipeline so the diff isn't that helpful. I could possibly move to a new pipeline coarsen_restarts2 to make the review process easier.
- Reliable I/O
  - Isolate I/O. Most of the functions are now pure functions.
  - Use battle-tested I/O tools. `vcm.open_tiles` for lazily loading the input data, `WriteNetCDFs` for saving the netCDF outputs.
  - remove I/O code from `vcm.coarsen` 
- Avoid out of memory errors w/o an on-disk cache by reduce each parallel problem size by parallelizing over times *and* tiles. With this change the problem fits into memory, so no local caching is needed.

- change APIs of `vcm.coarsen_restarts_on_pressure` and `vcm.coarsen_restarts_on_sigma` so they are pure.

# Other changes
- Add a routine `synth.generate_restart_data` for making fake restart data.
- Refactor `FunctionSource` to `fv3net.pipelines.common`


# TODO 

- [x] Check that restart files work with one-step jobs.

VCMML-308 #resolve