Allow piggy-backed prognostic runs (i.e. diagnostic-only ML) (#367)
Adds a command-line argument to the prognostic run's submission script to do "diagnostic_ml" i.e. compute the ML predictions but not apply them to the model state.

The baseline run in the end-to-end workflow is updated to use this option instead of not passing an ML model. This ensures that the baseline run outputs will have the same format as the prognostic run outputs. It will also provide useful diagnostic information. The option to not pass an ML model url (and thus use the default runfile) is maintained.

For piggy-backed runs, the diagnostic predictions of <dQ1> and <dQ2> are output under the names "net_heating_diagnostic" and "net_moistening_diagnostic". Piggy-backed runs still output variables called "net_heating" and "net_moistening" but they are zero-valued.