Add option to not raise error for crashed prog run (#395)
We sometimes expect prognostic or baseline runs to crash but still want subsequent steps such as post-processing to be performed by the end-to-end orchestrator.

This adds an option to not raise an error if the prognostic run crashes. Default behavior remains the same (so, integration test will still fail if prognostic run crashes).