Post process prognostic run automatically (#406)
This PR saves a few file-transfers with GCS by post-processing the files within the pod. This required orchestrating several containers, something that `run_kubernetes` was not designed to do. Instead of returning one-monolothic k8s job, this PR introduces several composable functions for creating different containers for the different stages of the prognostic run. This enables a better separation of concerns between running the model and the I/O heavy stages, and could be reused by other fv3 running applications.

Changes include
- A new module `fv3kube.containers` containing some composable functions for creating and modifying kubernetes containers. There are seperate containers for writing the run directory, running the model, and post processing the rundirs. In particular, the fv3 container no longer interacts with cloud storage or uses fv3run..
- `orchestrate_submit_job.py` no longer interacts with GCS. All configurations are passed via k8s python api.
- hard code the diag_table and the runfile. These configurations were helpful for debugging but increased the configuration burden on the user, and made our prognostic runs pretty homogeneous. The newly added `make test_prognostic_run` in fv3net root is an alternative for iterative development. (I'm open to restoring this options)
- Modify the post_processing script to make it more testable, and automatically copy all unprocessed files (except for broken symlinks).