Still upload failed prognostic runs
Because the prognostic run uses initContainers to separate the fv3 run
from the post processing, it will not upload the data if the fv3
container has an exit code other than 0. This commit ensures that even
failed runs with return 0, allowing the upload to proceed. The return
code is saved to a file.