Workflow for ml diags from mappers
Creates a workflow offline_ml_diags for calculating diagnostic quantities related to ML model predictions. This initial PR uses a similar structure as the training data diags workflow, with a few changes to accomodate the ML predictions. The outputs of this workflow are a netcdf of the quantities calculated by the diagnostics_utils.reduce_to_diagnostic: conditional time avg's and vertical integrals of ML predicted dQ1, dQ2.