Fix image-tag typo, add baseline run for one-step (#20)
Fix a bug in the one-step end-to-end yaml (image_tag should be image-tag) and add a baseline (no ML) run for the one-step jobs. This way we will have baseline runs for physics-off and clouds-off from the one-step, while the nudging workflow does a baseline run for a physics-on configuration.