Make metrics calc robust to short prognostic runs (#424)
Currently the metrics calculation as part of the prognostic run report assumes that all prognostic runs are at least 3 days long. This PR fixes that by returning an empty dataset for 3-day metrics (drift and rmse) if a run is too short.

Also fixes the prognostic run post-processing script to make sure it saves zarr stores with consolidated metadata. This used to be the case (and prognostic run report assumes it) but this behavior changed in #406.

* Save zarrs with consolidated metadata for prog run
* Return empty dataset for 3-day metrics on IndexError