Allow use of dataflow.sh script in e2e orchestrator (#428)
To allow use of the standardized dataflow submission script in e2e jobs, don't have the e2e orchestrator add extra_package or setup_file arguments. These will be added by the submission script.

The integration test still uses the direct runner for the create_training_data step, since I found in #372 that this is much faster for a small set of timestamps. But this PR will allow production workflows to use the dataflow runner.