Reduce regression datasets to C8, add synth-based grid dataset (#435)
Training unit tests were taking prohibitively long on my machine, so I reduced the resolution of the synth-generated mappers from C48 to C8 with 19 vertical levels instead of 79. I similarly updated the regression output targets.

The grid_dataset fixture is also updated to be synth-based. Only land-sea mask is set to reasonable values (0/1/2) as is needed to make the existing tests pass. Area is also confined to the range 1-2 for future-proofing.