modify ML sample datasets to use sample as first dimension (#440)
I was writing a test to ensure that the new Keras models produce datasets which have the same metadata as the "target" batch data, and the data wasn't matching because the "sample" dimension was last in the batch data. The ordering either way doesn't affect training, because the pack routines automatically move the sample dimension to the first dimension. Because it makes some sense to have the training arrays in the batch dataset use the same convention as when training occurs, I've transposed the batch arrays so that the sample dimension is first.

I also removed an unnecessary eager load in the batch dataset creation routine, which may avoid downloading data twice when generating arrays with just inputs and just outputs from a batch.