Add option to output only a subset of nudging time-steps (#438)
The nudging run produces over a million files totally over 1 TB of output.  So far, we have only used a small fraction of this data for training/analysis. This PR adds option to output only a subset of time-steps for the nudging run. While there are other options to reduce the size sub-setting in time seemed like an easy option that is compatible with the current loaders.

This is optional, so still possible to output the full dataset.

Partially solves #437

- [x] Blocked by #436