[VCMML-368] Create movie as part of prognostic run report (#431)
Add script for creating movies from prognostic run output. Currently, one movie (with plots of column heating/moistening) is created, but the script is written to be easily extensible.

The script is plugged into the argo workflow, and runs after the report generation. It takes about 15 minutes to generate the heating/moistening movie (making maps is fairly slow).