Refactor nudging fixture implementation
The fixtures contain some logic that we might want to use outside of a
fixture. For example, to combine mapper fixtures, we would first need to
generate their directory structure. Currently, the fixtures assume that
each dataset is mostly independent.

This PR refactors the logic for generating the fine res and nudging
datasets, so that it can be used without the fixtures.