[VCMML-382] Add offset_seconds argument to nudging mapper (#460)
When the nudging run saves subselected timesteps, the nudging mapper interprets the times as 1 timestep behind the time that was selected. This should probably be addressed by the subsetting/stage monitor in the nudging runfile, but the fine resolution mapper handles a similar problem using its `offset_seconds` argument. 

This PR adds an offset_seconds argument to the nudging mapper. It refactors the corresponding logic from the `loaders.mappers._fine_resolution` into a pair of general purpose transformations in `loaders.mappers._transformations`. Time time offsetting logic is moved to `vcm.convenience` which contains most of our time-related code.