Fix multithreaded movie still generation
The movie generation script seemed to consistently hang and have slow
performance, despite the succesful initial implementation by @oliwm.

This is likely related to a dask threadpool clashing with
multiprocessing. See this article:
https://pythonspeed.com/articles/python-multiprocessing/

This implements the change suggested by that article and also uses the
single-threaded dask scheduler (which has worked for similar issues in
dataflow).

Resolves #457