Added initial neural network training entrypoint (#443)
This PR creates an entrypoint for training a basic dense neural network, following the style of the sklearn training code.

The training is not hooked into the integration tests, and the network has not been tested for performance or optimized. I'm cutting the PR off here to keep the PR size reasonable.

The training should be hot-swappable for sklearn training, provided the configuration is appropriate (e.g. using `DenseModel` as the model type with appropriate or no hyperparameters).

Normalization will need a more sophisticated solution, right now batch normalization is applied to the inputs and no normalization is applied to the outputs. This could be done compositionally using a similar approach to `ArrayPacker`, or we could manually define frozen input and output layers which apply the normalization.

Documentation in this PR is mainly docstrings and unit/integration tests. Let me know if you'd like any additional documentation.

Tests are included both for the external `python -m` entrypoint and for the underlying routine. Serialization is tested so that the model result is unchanged when saving and loading against the filesystem.

Unfortunately save and load routines must take in directory paths and not file objects, as the keras routines do so. We need to wrap `load` and `dump` filesystem operations in temporary local directories which are copied to/from the provided path using `fsspec`, so that remote locations are usable. This is done in a follow-on PR #444 .