[VCMML-383] Nudging fine-res hybrid run (#28)
This PR contains an argo workflow for the end-to-end hybrid nudging/fine-res workflow.

The general purpose workflow templates in the `workflow-templates` folder could possible be moved into fv3net and referenced here. The nudging workflow-template is likely particularly useful and could be easily reused in other workflows. The Sklearn training template is somewhat bespoke for this workflow.

I have tested these on a 1 hour trial nudged run.

The submission templates the configurations using `yq`, `jq`, and bash. These are pretty lightweight requirements, but I would be open to containerizing them as well if we really want to avoid this. Another alternative would be to check the final `config.json` file into version control.

- [x] Switch all the configurations to "production" settings.