[VCMML-392] Remove nans in nudged mapper (#471)
The nudged run does not set the correct fill_value, leading to
NaNs when in the zarr data is loaded by xarray.

This is being addressed in this PR
https://github.com/VulcanClimateModeling/fv3gfs-python/pull/85

But, for now, we want to be able to train the ML with additional
predictors off existing runs. This quick hack will allow this.