Add workflow to do nudged-to-obs run with checkpointing (#465)
Add workflow to do run that is nudged towards observations and saves checkpointed state. (Actually, this workflow can just as easily do non-nudged runs, since that's just a configuration option).

In order to allow use of nudge-to-obs run output by ML and increase reliability of the upload step, this PR also refactors the post-processing script to be able to run on more general outputs (with user-specified chunking) and use a docker image separate from the fv3net image. The prognostic run submission is modified to use this new docker image.