Refactor time loop of prognostic runfile (#493)
The sklearn_runfile time loop is getting complicated and combines time-stepping with I/O logic. Any modifications required manually patching code into the time-loop, which violates the [open/closed principle](https://en.wikipedia.org/wiki/Open%E2%80%93closed_principle), and could lead to spaghetti code.

This PR refactors the actual time-stepping logic to a new class, which should be somewhat easier to extend. The I/O logic is a separate concern so I left it in `__main__`.

I don't think this is final state of this refactor but I think it is a positive improvement.

Other changes:
- the tests now verify that the checksum of one of the restart files remains unchanged. This makes the refactor easier.