Consolidate nudging into prognostic run folder (#495)
This PR moves the nudging workflow contents into the prognostic_c48_run workflow directory in anticipation of merging some of the workflows which perform a flavor of "run fv3gfs". I kept the nudging in a separate subdirectory to not interfere with others' edits. Additionally, this PR adds a nudging regression test that checks the basic operation of the nudging runfile.

I verified that make run still works from the nudging folder and that using the argo workflow template completes successfully.

* Start of nudging regression

* Add configuration fixuture

* Unique nudging regression name

* Linting

* Change nudging to use gcs paths

* Move nudged into prognostic directory

* Makefile works

* Gets to test_nudge_run and fails

* Passes nudged regression test

* Linting

* Comments

* Remove workflows/nudging from mypy

* More cleanup

* Move nudged regression into test_regression for prognostic_c48_run

* Finish regression test merge

* Update fv3net image to current for argo workflow template

* Change restart_dir fixture return

* Naming clarification

* Linting

* Address OliWM's comments

* Update nudging regression skip condition

* Linting

* Bump checksum

* Re-remove nudging from type checks after master merge

* Warn on checksum failure instead of failing