Load external grid spec for C384 diagnostic coarsening (#503)
Currently the C384 SHiELD diagnostics coarsening workflow assumes that the area variable is within the dataset being coarsened. That is not always true (e.g. for the atmos_8xdaily outputs) so this PR modifies the workflow to load in a separate C384 grid spec and use the area from there. It also removes the assumption that the output diagnostics should be saved as gfsphysics_15min_coarse.zarr.