Fix master by using master of fv3util
The test of the prognostic run file was detecting nans in the saved
data. This problem was fixed upstream in fv3util, but there is a lot of
red-tape involved in updating the version of fv3util built into the
fv3gfs-python image. According to the current release instructions of
fv3gfs-python, this is the necessary workflow:
1. Fix/improve fv3util with PR
2. Bump version of fv3util
3. make a PR to fv3gfs-python pointing the submodule to this new version
4. bump version of fv3util
5. make a PR to fv3net using this new version in the prognostic run
   dockerfile

Manually patching a custom fv3util is easier, but potentially more risky
since it could lead to version incompatibilities