Remove failing nan check
This check was added in the last PR, but is not currently passable by
the ZarrMonitor. The fix is upstream, but it will take at least a day to
propagate in this repo.

This new test would not have passed anyway on the old code anyway,
because it used zarr.create_dataset also, which uses a default
fill_value of 0.