Isolate the test/run environment of the prognostic run (#533)
The prognostic run is the only workflow that we could not effectively test in the "fv3net" global environment. For this reason, we refactored some of it's pure python components to a external package "runtime" that could be tested in the default "fv3net" environment. This package is only used by the prognostic run. This stop-gap measure gave us some coverage, but led to bloat in the dependencies of the global "fv3net" environment (namely adding fv3util as a submodule of this repository and installing it in the default environment that all workflows use). Runtime was also not really tested in the actual setting it is used (the prognostic_run docker image).

This PR completely isolates the testing, deployed, and development environment of the prognostic_c48_run. The runtime folder is refactored to `workflows/prognostic_c48_run` and now tested within the docker image.

- Add a new entrypoint script which installs any packages within the `/external/` folder of the docker image. This is used for developing using bind-mounts.
- Add tests in a dockerfile stage which won't rerun if no code changes.
- Modify the `workflows/prognostic_c48_run` dev environment to use docker-compose and simplify the makefile.
- Disable unit tests of the one_step_job runfile. This workflow will be deprecated so is not worth refactoring.