Add mappers and opening functions for nudge-to-obs runs (#534)
Custom mappers are required to load data from "nudge to obs" runs because the Fortran code does the nudging within the "step_physics" fv3gfs-python routine. Thus, a tendency computed by taking the difference between the before and after physics states is actually the sum of the physics tendency and the nudging tendency. So to compute the pQs it is necessary to subtract the nudging tendency. Also, to get the state of the model just before the nudging tendency is computed and applied, it is necessary to subtract the nudging tendency times the physics timestep from the "after physics" (i.e. "after nudging") state.

New public API:
- Adds two new public functions to loaders package: open_merged_nudge_to_obs and open_merged_nudge_to_obs_full_tendencies.

New internals:
- SubtractNudgingIncrement and SubtractNudgingTendency classes in loaders/mappers/_nudged.py

I added this code to the existing _nudged.py module in order to leverage pre-existing mappers and test infrastructure.